added more in delete.php

+++ app/lists/delete.php
@@ -4,11 +4,19 @@ declare(strict_types=1);
 
 require __DIR__ . '/../autoload.php';
 
-if (isset($_POST['id'])) {
-    $title = filter_var($_POST['id'], FILTER_SANITIZE_NUMBER_INT);
-    $statement = $database->prepare('DELETE FROM lists WHERE id = :id');
-    $statement->bindParam(':id', $title, PDO::PARAM_INT);
+if (isset($_GET['id'])) {
+    $id = $_GET['id'];
+    $user_id = $_SESSION['user']['id'];
+
+    $statement = $database->prepare('DELETE FROM lists WHERE user_id = :user_id AND id = :id');
+    $statement->bindParam(':user_id', $user_id, PDO::PARAM_INT);
+    $statement->bindParam(':id', $id, PDO::PARAM_INT);
+    $statement->execute();
+
+    $statement = $database->prepare('DELETE FROM tasks WHERE user_id = :user_id AND list_id = :id');
+    $statement->bindParam(':user_id', $user_id, PDO::PARAM_INT);
+    $statement->bindParam(':id', $id, PDO::PARAM_INT);
     $statement->execute();
 }
 
-redirect('/');
+redirect('/lists.php');
